requirements.txt
django
djangorestframework
django-phonenumber-field
django-cors-headers